fix: 🐛 still unfixed: templates inside build unreachable
templates folder seems to either get compiled and then work, but it
gives us compiled files, (as in built) or manually put in the build
folder - and then nothing gets copied. there is a dev branch that shows
this difference. this is due to tsconfig.json where ./src/templates/**/*
is an excluded folder.