feat: work on the Create a custom type feature
add some documentation, add validation for all inputs of the process, add some TODOs. add general
workflow that leads up to the creation of the actual custom type: from the initial naming of the
custom type, to the collecting of props, to validating the whole thing one last time and creating
it. Validation for props is almost complete: handleDuplicateKeysInCustomType(), at
validation.util.ts, still needs to be written.